update : optimize code  #7

+++ pose-management-api/src/main/java/com/zlht/pose/management/api/service/impl/UserServicesImpl.java
@@ -2,17 +2,12 @@ package com.zlht.pose.management.api.service.impl;
 
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-
-
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-
-
+import com.zlht.pose.management.api.enums.Status;
 import com.zlht.pose.management.api.service.UserServicesI;
 import com.zlht.pose.management.api.utils.Result;
 import com.zlht.pose.management.dao.entity.User;
 import com.zlht.pose.management.dao.mapper.UserMapper;
-
-
 import com.zlht.pose.management.tools.service.Argon2PasswordEncoder;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
@@ -20,8 +15,8 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 @Service
 public class UserServicesImpl extends BaseServiceImpl<User> implements UserServicesI {
@@ -32,74 +27,88 @@ public class UserServicesImpl extends BaseServiceImpl<User> implements UserServi
     UserMapper userMapper;
 
     @Override
-    public Result<User> queryUserList(int type, int pageNum, int pageSize, String nickname) {
+    public Result<User> queryUserList(int type, int pageNum, int pageSize, String keyword) {
 
-
-        List<User> userList = new ArrayList<>();
+        Result result = new Result();
         Page<User> page = new Page<>(pageNum, pageSize);
 
         QueryWrapper<User> wapper = new QueryWrapper<User>();
         if (type != -1) wapper.eq("type", type);
-        if (nickname != null) {
-            wapper.like("nickname", nickname);
+        if (keyword != null) {
+            wapper.like("nickname", keyword);
         }
         Page<User> userPage = userMapper.selectPage(page, wapper);
-        if (userPage != null) {
-            for (User user : userPage.getRecords()) {
-                userList.add(user);
-            }
-            return success(userList);
-        } else {
-            return faild(400, "未查询到用户！");
-        }
+        result.setCode(Status.SUCCESS.getCode());
+        result.setMsg(Status.SUCCESS.getMsg());
+        result.setData(userPage.getRecords());
+        return result;
     }
 
     @Override
-    public Result<User> createUser(User user) {
+    public Map<String, Object> createUser(User user) {
 
-        if (!validateUserName(user)) {
-            return faild(400, "用户名或昵称不符合规范！");
+        Map<String, Object> map = new HashMap<>();
+        if (checkUserExistByName(user)) {
+            putMsg(map, 400, "用户已存在！");
+            return map;
         }
-        //exist?
-        if (checkUserExist(user)) {
-            return faild(400, "用户名重复！");
+        if (!validateUserName(user)) {
+            putMsg(map, 400, "用户名或昵称不符合规范！");
+            return map;
         }
         //管理员的话加密 密码
 
         int resNum = userMapper.insert(encipher(user));
         if (resNum >= 1) {
-            return success(null);
+            putMsg(map, Status.SUCCESS.getCode(), "新建用户成功！");
         } else {
-            return faild(400, "插入用户失败");
+            putMsg(map, 400, "新建用户失败！");
         }
-
+        return map;
     }
 
 
     @Override
-    public Result<User> updateUser(int id, User user) {
+    public Map<String, Object> updateUser(int id, User user) {
+
+        Map<String, Object> map = new HashMap<>();
+        if (!checkUserExistById(id)) {
+            putMsg(map, 400, "所更新用户不存在！");
+            return map;
+        }
+
+        if (checkUserExistByName(user)) {
+            putMsg(map, 400, "用户已存在！");
+            return map;
+        }
 
         QueryWrapper queryWrapper = new QueryWrapper();
         queryWrapper.eq("id", id);
         int update = userMapper.update(encipher(user), queryWrapper);
         if (update >= 1) {
-            return success(null);
+            putMsg(map, Status.SUCCESS.getCode(), "更新用户成功！");
         } else {
-            return faild(400, "更新用户失败");
+            putMsg(map, 400, "更新用户失败！");
         }
+        return map;
     }
 
     @Override
-    public Result<User> deleteUser(int id) {
+    public Map<String, Object> deleteUser(int id) {
+        Map<String, Object> map = new HashMap<>();
+        if (!checkUserExistById(id)) {
+            putMsg(map, 400, "所更新用户不存在！");
+            return map;
+        }
         QueryWrapper queryWrapper = new QueryWrapper();
         queryWrapper.eq("id", id);
         int delete = userMapper.delete(queryWrapper);
         if (delete >= 1) {
-            return success(null);
+            putMsg(map, Status.SUCCESS.getCode(), "删除用户成功！");
         } else {
-            return faild(400, "删除用户失败！");
+            putMsg(map, 400, "删除用户失败！");
         }
-
+        return map;
     }
 
     /**
@@ -109,23 +118,30 @@ public class UserServicesImpl extends BaseServiceImpl<User> implements UserServi
      * @param password
      * @return
      */
-    public Result authorizedUser(String username, String password) {
-
-        if (username == null) return faild(400, "请输入用户名！");
-        if (password == null) return faild(400, "请输入密码！");
-
+    public Map<String, Object> authorizedUser(String username, String password) {
+        Map<String, Object> map = new HashMap<>();
+        if (username == null) {
+            putMsg(map, 400, "请输入用户名！");
+            return map;
+        }
+        if (password == null) {
+            putMsg(map, 400, "请输入密码！");
+            return map;
+        }
         User user = userMapper.queryUserByUserName(username);
         if (user != null) {
             String encipherPassword = user.getPassword();
             boolean check = Argon2PasswordEncoder.matches(encipherPassword, password);
             if (check) {
-                return success(null);
+                putMsg(map, Status.SUCCESS.getCode(), "登录成功！");
             } else {
-                return faild(400, "登录失败！");
+                putMsg(map, 400, "用户名或密码错误！");
+                return map;
             }
         } else {
-            return faild(400, "登录失败！");
+            putMsg(map, 400, "用户名或密码错误！");
         }
+        return map;
     }
 
     /**
@@ -135,9 +151,22 @@ public class UserServicesImpl extends BaseServiceImpl<User> implements UserServi
      * @return
      */
     @Override
-    public boolean checkUserExist(User user) {
+    public boolean checkUserExistByName(User user) {
         QueryWrapper queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("username", user.getUsername());
+        return userMapper.exists(queryWrapper);
+    }
+
+    /**
+     * 检查用户是否重复
+     *
+     * @param id
+     * @return
+     */
+    @Override
+    public boolean checkUserExistById(int id) {
+        QueryWrapper queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("id", id);
         return userMapper.exists(queryWrapper);
     }
 
@@ -166,7 +195,6 @@ public class UserServicesImpl extends BaseServiceImpl<User> implements UserServi
                 || StringUtils.containsWhitespace(username) || StringUtils.containsWhitespace(nickname)) {
             return false;
         }
-
         return true;
     }
 }
